# soul-shield/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "date না দিলে আজকের তারিখ ধরা হবে। যেই দিনে recurrence_days ম্যাচ করবে সেই task গুলোই দেখাবে।",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "List today's (or given date's) tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, default: today",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/task.TaskWithStatusResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User বা Admin task তৈরি করতে পারবে। is_global=true শুধু admin ব্যবহার করতে পারবে।\n▹ for recurrence_days: 0=Sunday ... 6=Saturday.\n▹ recurrence_type: 'daily', 'weekly', 'custom'. 'daily' হলে সব দিন auto বসিয়ে দেওয়া হবে insert time এ\n▹ category_id: এই field না দিলে global task হিসেবে create হবে, category_id রাখতে চাইলে category id must দিতে হবে\n▹ task_type \"normal\" হলে target_count optional, আর \"counter\" হলে target_count must দিতে হবে\n▹ \"counter\" type task for dhikr",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "Task Info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/task.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tasks/{taskId}/subtasks/{subTaskId}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "date না দিলে আজকের তারিখ ধরা হবে। Sub-task এর নিজের recurrence নাই - parent task যেই দিন scheduled সেই দিনেই complete করা যাবে।",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark a normal-type sub-task as completed for a given date",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent Task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sub-Task ID",
                        "name": "subTaskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional date",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/task.CompleteTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.SubTaskCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}/subtasks/{subTaskId}/increment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "amount ব্যাচ আকারে পাঠান। Sub-task এর নিজের recurrence নাই - parent task যেই দিন scheduled সেই দিনেই increment করা যাবে।",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Increment progress of a counter-type sub-task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent Task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sub-Task ID",
                        "name": "subTaskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Increment amount",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.IncrementTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/task.SubTaskCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/task.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "task.CompleteTaskRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "না দিলে আজকের তারিখ ধরা হবে; format: YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-07-04"
                }
            }
        },
...
        "task.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Read at least 1 page daily"
                },
                "is_global": {
                    "type": "boolean",
                    "example": false
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        3,
                        5
                    ]
                },
                "recurrence_type": {
                    "type": "string",
                    "enum": [
                        "daily",
                        "weekly",
                        "custom"
                    ],
                    "example": "custom"
                },
                "reminder_time": {
                    "description": "optional, \"HH:MM\" 24-hour format",
                    "type": "string",
                    "example": "18:30"
                },
                "reward_text": {
                    "type": "string",
                    "example": "Alhamdulillah! +1 for jannah"
                },
                "sub_tasks": {
                    "description": "SubTasks দিলে parent task এর status এখন থেকে sub-task completion থেকে derive হবে\n(pending/partially_completed/completed) - parent নিজে সরাসরি complete করা যাবে না।",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskInput"
                    }
                },
                "target_count": {
                    "description": "task_type=counter হলে required",
                    "type": "integer",
                    "example": 100
                },
                "task_type": {
                    "type": "string",
                    "enum": [
                        "normal",
                        "counter"
                    ],
                    "example": "normal"
                },
                "title": {
                    "type": "string",
                    "example": "Read Quran"
                }
            }
        },
        "task.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "task.IncrementTaskRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "batch amount, single-tap না",
                    "type": "integer",
                    "example": 10
                },
                "date": {
                    "type": "string",
                    "example": "2026-07-04"
                }
            }
        },
        "task.SubTaskCompletionResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string",
                    "example": "2026-07-04"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "parent_reward_text": {
                    "type": "string"
                },
                "parent_status": {
                    "type": "string",
                    "example": "partially_completed"
                },
                "parent_task_id": {
                    "type": "integer",
                    "example": 1
                },
                "progress_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "sub_task_id": {
                    "type": "integer",
                    "example": 5
                },
                "sub_task_title": {
                    "type": "string",
                    "example": "Read 1 page"
                }
            }
        },
        "task.SubTaskInput": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "update এ existing sub-task ম্যাচ করাতে; create এ omit করবেন",
                    "type": "integer",
                    "example": 5
                },
                "target_count": {
                    "description": "task_type=counter হলে required",
                    "type": "integer",
                    "example": 10
                },
                "task_type": {
                    "type": "string",
                    "enum": [
                        "normal",
                        "counter"
                    ],
                    "example": "normal"
                },
                "title": {
                    "type": "string",
                    "example": "Read 1 page"
                }
            }
        },
        "task.SubTaskStatusResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "progress_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "sub_task_id": {
                    "type": "integer",
                    "example": 5
                },
                "target_count": {
                    "type": "integer",
                    "example": 10
                },
                "task_type": {
                    "type": "string",
                    "example": "normal"
                },
                "title": {
                    "type": "string",
                    "example": "Read 1 page"
                }
            }
        },
...
        "task.TaskResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Read at least 1 page daily"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_global": {
                    "type": "boolean",
                    "example": false
                },
                "owner_id": {
                    "type": "integer",
                    "example": 3
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        3,
                        5
                    ]
                },
                "recurrence_type": {
                    "type": "string",
                    "example": "custom"
                },
                "sub_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskStatusResponse"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Read Quran"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "task.TaskWithStatusResponse": {
            "type": "object",
            "properties": {
                "category_color": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_global": {
                    "type": "boolean"
                },
                "progress_count": {
                    "type": "integer"
                },
                "recurrence_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "recurrence_type": {
                    "type": "string"
                },
                "reminder_time": {
                    "type": "string"
                },
                "reward_text": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "sub_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/task.SubTaskStatusResponse"
                    }
                },
                "target_count": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "task_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}